+++ R/dengue_MS.R
@@ -1,12 +1,15 @@
 #' Dengue cases in *Mato Grosso do Sul*
 #'
-#' Data frame containing the monthly number of notified dengue cases by
-#' municipality in the *Mato Grosso do Sul* state of Brazil, as well as set 
-#' of spatial and spatio-temporal covariates
-#' (e.g., environmental, socio-economic and meteo-climatic factors).
-#' This dataset is a subset of the original containing observations
-#' over the entire Brazil. 
-#'
+#' Monthly number of notified dengue cases by
+#' municipality in the *Mato Grosso do Sul* state of Brazil and a set 
+#' of spatial and spatio-temporal covariates.
+#' 
+#' @details
+#' In addition to the dengue counts, the dataset contains a set of environmental,
+#' socio-economic and meteo-climatic factors. This dataset is a subset of the 
+#' original containing observations over the entire Brazil. 
+#' 
+#' @docType data
 #' @format 
 #' A data frame with 2,640 rows and 27 columns:
 #' \describe{
@@ -54,5 +57,5 @@
 #'   \item{date}{First day of the month in date format ("%d-%m-%Y").}
 #' }
 #' @source <https://github.com/drrachellowe/hydromet_dengue>
-
+#' @usage data(dengue_MS)
 "dengue_MS"