+++ R/map_MS.R
@@ -2,12 +2,14 @@
 #'
 #' Administrative boundaries (polygon geometries) of the 11 municipalities
 #' in *Mato Grosso do Sul* (Brazil) to be used with the `dengue_MS` data set.
-#'
+#' 
+#' @docType data
 #' @format 
 #' A simple feature (sf) object including 11 rows and 2 columns:
 #' \describe{
 #'   \item{code}{Unique ID number to each micro region (11 units).}
 #'   \item{geometry}{Geometries of the sf multipolygon.}
 #'   }
-#'
+#' 
+#' @usage data(map_MS)
 "map_MS"