+++ NEWS.md
@@ -5,4 +5,10 @@
 # GHRexplore 0.1.1
 
 * Fixed tests failing with cowplot 1.2.0.
-* Fixed minor bug in plot_map.
+* Fixed minor bug in `plot_map()`.
+
+# GHRexplore 0.2.0
+
+* New plotting function: `plot_timeseries2()` for dual-axis time series plots.
+* Increased flexibility of `plot_map()` (line colour, width, alpha).
+* *by_year* argument deprecated in `plot_map()` and substituted by *aggregate_time*.